pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "fairly"
version = "0.1.0a1"
authors = [
  { name="Serkan Girgin", email="s.girgin@utwente.nl" },
  { name="Manuel Garcia Alvarez", email="m.g.garciaalvarez@tudelft.nl"},
  { name="Jose Urra Llanusa", email="j.c.urrallanusa@tudelft.nl"},
]
description = "A package to create, publish, and download research datasets"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "requests",
	"pyyaml",
]

[project.urls]
"Homepage" = "https://github.com/ITC-CRIB/fairly"
"Bug Tracker" = "https://github.com/ITC-CRIB/fairly/issues"